Implement line functionality
The line class for now compiles its own shader, which giving no
access or option to change it to outside classes. Nearly all of the
variables are bound to an IDContainer object, a shared pointer of
which more pointers are created when a line object is copied. The only
variable that is unique to the object is the color.

There is a problem right now upon compilation, with the glm forward declaration
header, as it fails upon reaching the unique_ptr in the header
file. As a temporary workaround, the proper glm header is used
instead. This is of course unnaceptable and will be changed as soon as possible.