Update line class
The color field has been made to affect all instances of a line
class. This with the intention of removing issues concerning the use
forward declarations of glm classes in the header file. The unproper
use of glm.hpp in the header file was removed, replace with the use of
the proper forward declaration fwd.hpp.

The copy constructor and the copy operator have been removed, as they
are no longer needed, as the default c++ generated ones will now be enough.