Add load_points function with vec3 array as parameter
In order to avoid the inefficiencies of the vector object, a function
has been added in order to allow the line class to load points into
its VBO through a vec3 array, as well as a size parameter. The
underlying implementation has also been changed to also use this even
when a vector is passed.