Add automatic demotion of OpenGL version
A new method has been added to the initialization of the program which
iterates through the available OpenGL versions, starting from 4.3, and
works backward until reaching the maximum version of OpenGL supported
on the system.

This is done by creating a GLFW version with a specified version. If
this fails, then the version must not be supported.